+++ ocaml-lsp-server/test/e2e-new/helpers.mli
@@ -1,0 +1,6 @@
+open Test.Import
+
+val uri : Uri.t
+
+val test :
+  ?extra_env:string list -> string -> (unit Client.t -> unit Fiber.t) -> unit